ai-modules/career-guidance-ai/requirements.txt
streamlit==1.28.1
google-generativeai==0.8.3
google-cloud-customsearch==2.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
requests==2.31.0
serpapi==2.0.0
